requirements.txt
streamlit
tensorflow
pillow
numpy
google-generativeai
scipy
fpdf2
